ISSUE #2548 - Refactor showing of new dialog
