+++ frontend/routes/components/dialogContainer/components/index.ts
@@ -3,5 +3,5 @@ export * from './errorDialog/errorDialog.component';
 export * from './loadingDialog/loadingDialog.component';
 export * from './newUpdateDialog/newUpdateDialog.component';
 export * from './suggestedTreatmentsDialog/suggestedTreatmentsDialog.component';
+export * from './redirectToTeamspaceDialog/redirectToTeamspaceDialog.component';
 export { default as RevisionsDialog } from './revisionsDialog/revisionsDialog.container';
-export { default as UnauthorizedAccessDialog } from './unauthorizedAccessDialog/unauthorizedAccessDialog.container';
